requirements.txt
google-generativeai>=0.8.0
vaderSentiment>=3.3.2
python-dotenv>=1.0.0